+++ Package.swift
@@ -9,8 +9,11 @@ let package = Package(
     products: [
         .library(name: "NetworkExtensions", targets: ["NetworkExtensions"]),
     ],
+    dependencies: [
+        .package(url: "https://github.com/WeTransfer/Mocker.git", from: "3.0.1")
+    ],
     targets: [
         .target(name: "NetworkExtensions", path: "Sources"),
-        .testTarget(name: "NetworkExtensionsTests", dependencies: ["NetworkExtensions"], path: "Tests")
+        .testTarget(name: "NetworkExtensionsTests", dependencies: ["NetworkExtensions", "Mocker"], path: "Tests")
     ]
 )